[+] Finally a good version of prompting. This version handle all of theses functions  - Typing characters  - Backspace character  - Multiline prompt  - CTRL A (go first character) -> handle multiline  - CTRL E (go last character) -> handle multiline  - CTRL K (cut from cursor) -> handle multiline  - CTRL Y (paste from cursor) -> handle multiline  - CTRL C (cancel the prompt) -> handle multiline  - Left ARROW (go to left for 1 character) -> handle multiline  - Right ARROW (go to right for 1 character) -> handle multiline  - History UP (go 1 history before) -> handle multiline history prompt  - History DOWN (go 1 history after) -> handle multiline history prompt
All of theses can be down in the same time, doesn't matter if you are at the top of the terminal, at the end of the terminal, if you have a small terminal or a big one.
I like this version and think this can be a final one.

If you see any bugs please make an issue!